Update service user deleteId, with spring 3+

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -38,9 +38,11 @@ public class UserService {
 
 	public void delete(Long id) {
 		try {
-			userRepository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
+			if (userRepository.existsById(id)) {
+				userRepository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
 		} catch (DataIntegrityViolationException e) {
 			throw new DataBaseException(e.getMessage());
 		}
